refactor: extracted lazy calculation of resolved subjects. Removed unused imports
